Refactoring code and merge duplicated CSS
