--- telos-frontend/src/components/habit-tracker/DashboardHabitTracker.module.css
@@ -1,166 +1,0 @@
-@charset "utf-8";
-/* CSS Document */
-* {
-  margin: 0;
-  padding: 0;
-}
-
-.hidden {
-  display: none !important;
-}
-
-.container {
-  position: absolute;
-  width: 562px;
-  height: 200px;
-  text-align: left !important;
-  background: #ffffff;
-  mix-blend-mode: normal;
-  border: 2px solid #7a00f4;
-  box-sizing: border-box;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 19px;
-}
-
-.HabitTitle {
-  display: block;
-  height: 20%;
-  padding: 10px;
-}
-
-.AddIcon {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-}
-
-.HabitContent {
-  height: 20%;
-  padding: 10px;
-}
-
-.WeekSchedule {
-  display: block;
-  height: 40%;
-  width: 100%;
-  margin: auto;
-}
-
-.WeekSchedule ul {
-  display: flex;
-  width: 100%;
-  list-style: none;
-  align-items: center;
-  justify-content: center;
-}
-
-.unselect {
-  border: 1px solid #7a00f4 !important;
-  background-color: #ffffff !important;
-  color: #7a00f4 !important;
-}
-
-.WeekSchedule ul li button {
-  display: inline-flex;
-  margin: 10px;
-  width: 40px;
-  height: 40px;
-  border: 1px solid #9933ff;
-  border-radius: 50%;
-  background-color: #7a00f4;
-  color: #ffffff;
-  align-items: center;
-  justify-content: center;
-  outline: none;
-  cursor: pointer;
-}
-
-.WeekSchedule ul li button:disabled,
-.disable {
-  border: 1px solid #b0adb3 !important;
-  background-color: #b0adb3 !important;
-  color: #ffffff !important;
-}
-
-.menuContainer {
-  display: block !important;
-  width: 562px !important;
-  height: 408px !important;
-}
-
-.menuHeader {
-  display: inline-flex !important;
-  width: 100% !important;
-  height: 16% !important;
-  color: #ffffff !important;
-  background: #210043 !important;
-  border: 2px solid #210043 !important;
-  box-sizing: border-box !important;
-  align-items: center !important;
-}
-
-.menuHeader h2 {
-  left: 50% !important;
-}
-
-.menuBody {
-  width: 100% !important;
-  height: 84% !important;
-
-  background: #ffffff !important;
-  mix-blend-mode: normal !important;
-  border: 2px solid #210043 !important;
-  box-sizing: border-box !important;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25) !important;
-}
-
-#saveButton {
-  width: 77px !important;
-  height: 22px !important;
-  left: 20px !important;
-  margin-right: 146px !important;
-  background: #ffffff !important;
-  border-radius: 19px !important;
-}
-
-#closeButton {
-  width: 20px !important;
-  min-width: 0 !important;
-  height: 20px !important;
-  left: 170px !important;
-  padding: 0 !important;
-  background: #ffffff !important;
-  border-radius: 50% !important;
-}
-
-.habitInfo {
-  display: block;
-}
-
-.habitInfo,
-.weekInfo,
-.timeSetting {
-  width: 96%;
-  padding: 0 2%;
-  display: block;
-}
-
-.habitName,
-.habitDetail,
-.startDate,
-.repeat,
-.endDate {
-  width: 100%;
-  height: 50px;
-  display: inline-flex;
-  align-items: center;
-}
-
-#name,
-#description,
-#repeatsOn,
-#start,
-#selectWeeks,
-#end {
-  padding-right: 6px;
-}
